feat(gpu-frustum-culling): switch glsl hashmap keys to uint64_t instead of uvec2 and enable 64-bit shader support

+++ VoxelEngine/gpu_cache_allocator.h
@@ -45,8 +45,8 @@ private:
     struct ObjectAllocation
     {
         unsigned int totalElementCount = 0;
-        uint16_t startPage = PageNode::NULL_PAGE;
-        uint16_t endPage = PageNode::NULL_PAGE;
+        uint32_t startPage = PageNode::NULL_PAGE;
+        uint32_t endPage = PageNode::NULL_PAGE;
 
         Policy<ObjectID>::Handle policyHandle;
